Add copy to clipboard to CopyButton
As outlined in the GitHub issue for this feature, there are three
different options available for adding this functionality in Gutenberg.

I have opted for the oldest (and now deprecated ClipboardButton
component).

This does come with it the disadvantage of a depreaction notice but that
seems a small price to pay for backwards compatibility.

Initially, I inteneded to use the most modern solution
useCopyToClipboard (adding in Gutenberg 10.3.0, WordPress core 5.8.x),
my thinking being that since we were importing this hook from the latest
version of the package, it would work despite running on an older
version of Gutenberg.

import { useCopyToClipboard } from '@wordpress/compose';

Nope - this fails with with the following error

TypeError
_wordpress_composer__WEBPACK_IMPORTED_MODULE_1__.useCopyToClipboard is
not a function

This only deepens my confusion around importing packages vs
destructuring them from the "wp" global as discussed in
https://salferrarello.com/wordpress-gutenberg-import-versus-destructuring-global-wp/

At this point I could add a conditional and provide two (or more) sets
of code to create this copy functionality to ensure I don't throw any
deprecation notices or I can use this older (but fully functional
technique) and be done with it.

Resolves #17